requirements.txt
Flask==0.10.1
Jinja2==2.8
Markdown==2.6.5
MarkupSafe==0.23
PyYAML==3.11
Pygments==2.0.2
Werkzeug==0.11.2
argparse==1.2.1
backports.ssl-match-hostname==3.5.0.1
docker-py==1.6.0
docopt==0.6.2
grip==4.0.0
itsdangerous==0.24
path-and-address==1.0.0
requests==2.8.1
six==1.10.0
websocket-client==0.34.0
wheel==0.24.0
wsgiref==0.1.2
